% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testhello.R
\name{hello3}
\alias{hello3}
\title{Test string passing in the 'hello person' c function}
\usage{
hello3(name)
}
\arguments{
\item{name}{The name of the person to say hello to}
}
\description{
Test string passing in the 'hello person' c function
}
\examples{
hello3("Simon")
}
\keyword{test}

