% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/isodists.R
\name{isodists}
\alias{isodists}
\title{Test the 'Rmain' q2e c function}
\usage{
isodists(pepfn)
}
\arguments{
\item{pepfn}{peptide file name}

\item{dfn}{data file name}

\item{paramfn}{parameter file name}
}
\description{
Test the 'Rmain' q2e c function
}
\examples{
testq2e()
}
\keyword{test}

