% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isodists.R
\name{q2e_isodists}
\alias{q2e_isodists}
\title{Wrapper for the 'R_iso_seq' q2e c function}
\usage{
q2e_isodists(sequence)
}
\arguments{
\item{seqeunce}{the amino acid sequence}
}
\description{
Wrapper for the 'R_iso_seq' q2e c function
}
\examples{
q2e_isodists("IGQPGAVGPAGIR")
}
\keyword{test}

