% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q2eplotseqpeaks.R
\name{q2e_plotseqpeaks}
\alias{q2e_plotseqpeaks}
\title{Plots a set of mass spec peaks as mass/probability values}
\usage{
q2e_plotseqpeaks(data, myxlim)
}
\arguments{
\item{data}{a structure holding a vector of mass/probability values}
}
\description{
Plots a set of mass spec peaks as mass/probability values
}
\examples{
q2e_plotseqpeaks()
}
\keyword{isodists}
\keyword{q2e}

