% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodists.R
\name{q2e_tpeaks}
\alias{q2e_tpeaks}
\title{Wrapper for the 'R_iso_seq' q2e c function, to get theoretical peaks from a sequence}
\usage{
q2e_tpeaks(sequence)
}
\arguments{
\item{seqeunce}{the amino acid sequence}
}
\description{
Wrapper for the 'R_iso_seq' q2e c function, to get theoretical peaks from a sequence
}
\examples{
q2e_tpeaks("IGQPGAVGPAGIR")
}
\keyword{test}

