% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/q2ecalc.R
\name{q2ecalc}
\alias{q2ecalc}
\title{Mimic the original q2e program (for testing the individual functions)}
\usage{
q2ecalc(datalist = NA, peplist = NA, paramfile = NA)
}
\arguments{
\item{data_filelist}{text file containing list of data input files}

\item{peptide_filelist}{text file containing list of peptides}

\item{parameter_file}{text file containing parameters}
}
\description{
Mimic the original q2e program (for testing the individual functions)
}
\examples{
q2ecalc(data_filelist,peptide_filelist,parameter_file)
}
\keyword{q2e}
\keyword{testing}

