% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/q2eloadfiles.R
\name{q2eloadfiles}
\alias{q2eloadfiles}
\title{Create a structure that holds a list of files to be processed}
\usage{
q2eloadfiles(filelistfile, verbose = F)
}
\arguments{
\item{filelistfile}{name of a file holding the list of files}
}
\description{
Create a structure that holds a list of files to be processed
}
\examples{
q2eloadfiles()
}
\keyword{filelist}
\keyword{q2e}

