% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/q2eloadpeptides.R
\name{q2eloadpeptides}
\alias{q2eloadpeptides}
\title{Create a structure that holds a list of files to be processed}
\usage{
q2eloadpeptides(listfile, verbose = F)
}
\arguments{
\item{listfile}{name of a file holding the list of files}
}
\description{
Create a structure that holds a list of files to be processed
}
\examples{
q2eloadpeptides()
}
\keyword{peptidelist}
\keyword{q2e}

