% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q2eparams.R
\name{q2eparams}
\alias{q2eparams}
\title{Create a structure that holds the q2e parameters}
\usage{
q2eparams(fitpeaks = 4, galim = 0.02, snrlim = 3, firstmass = 800,
  lastmass = 4000, csv = 0)
}
\arguments{
\item{fitpeaks}{The number of peaks}

\item{galim}{The limit value for the genetic algorithm}

\item{snrlim}{The limit value for the signal to noise ratio}

\item{firstmass}{The smallest mass value}

\item{lastmass}{The largest mass value}

\item{csv}{flag to state whether to produce an ouput in csv format}
}
\description{
Create a structure that holds the q2e parameters
}
\examples{
q2eparams()
}
\keyword{parameters}
\keyword{q2e}

