% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rq2e}
\alias{rq2e}
\title{Run q2e via datafiles as in the C++ code}
\usage{
rq2e(argVector)
}
\arguments{
\item{paramFile}{the parameter file}
}
\description{
Run q2e via datafiles as in the C++ code
}
