% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/q2e.R
\name{testq2e}
\alias{testq2e}
\title{Test the 'Rmain' q2e c function}
\usage{
testq2e(dfn, pepfn, paramfn)
}
\arguments{
\item{dfn}{data file name}

\item{pepfn}{peptide file name}

\item{paramfn}{parameter file name}
}
\description{
Test the 'Rmain' q2e c function
}
\examples{
testq2e()
}
\keyword{test}

