% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/testhello.R
\name{hello2}
\alias{hello2}
\title{Test the 'hello' c calling function}
\usage{
hello2(n)
}
\arguments{
\item{n}{The number of times to say hello}
}
\description{
Test the 'hello' c calling function
}
\examples{
hello2()
}
\keyword{test}

