% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodists.R
\name{isodists}
\alias{isodists}
\title{Test the 'R_iso_main' q2e c function}
\usage{
isodists(pepfn)
}
\arguments{
\item{pepfn}{peptide file name}
}
\description{
Test the 'R_iso_main' q2e c function
}
\examples{
isodists("../testdata/cowPeptides")
}
\keyword{test}

